requirements.txt
streamlit
torch
transformers
Pillow
pytesseract
